Refactor/domain package (#5)
* feat: 조건에 맞는 전체 일정 조회 기능 구현
* refactor: 패키지 구조 정리
