+++ src/main/java/com/example/api/ScheduleApiApplication.java
@@ -1,0 +1,13 @@
+package com.example.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ScheduleApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ScheduleApiApplication.class, args);
+    }
+
+}
