+++ src/main/java/com/example/api/user/dto/UserRequestDto.java
@@ -1,0 +1,13 @@
+package com.example.api.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserRequestDto {
+    private String name; //회원명
+}
