+++ src/main/java/com/example/api/user/entity/User.java
@@ -1,0 +1,13 @@
+package com.example.api.user.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+    private Long userId; //회원 고유 id
+    private String name; //회원명
+}
