+++ src/main/java/com/example/api/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.example.api.user.repository;
+
+import com.example.api.user.entity.User;
+
+public interface UserRepository {
+    Long save(User user);
+}
