--- src/main/java/com/example/scheduleapi/ScheduleApiApplication.java
@@ -1,13 +1,0 @@
-package com.example.scheduleapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ScheduleApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ScheduleApiApplication.class, args);
-    }
-
-}
