--- src/main/java/com/example/scheduleapi/domain/User.java
@@ -1,13 +1,0 @@
-package com.example.scheduleapi.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-    private Long userId; //회원 고유 id
-    private String name; //회원명
-}
