--- src/main/java/com/example/scheduleapi/dto/user/UserRequestDto.java
@@ -1,13 +1,0 @@
-package com.example.scheduleapi.dto.user;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserRequestDto {
-    private String name; //회원명
-}
