--- src/main/java/com/example/scheduleapi/dto/user/UserResponseDto.java
@@ -1,11 +1,0 @@
-package com.example.scheduleapi.dto.user;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Getter
-public class UserResponseDto {
-    private Long userId; //회원 고유 id
-    private String name; //회원명
-}
