--- src/main/java/com/example/scheduleapi/repository/user/UserRepository.java
@@ -1,7 +1,0 @@
-package com.example.scheduleapi.repository.user;
-
-import com.example.scheduleapi.domain.User;
-
-public interface UserRepository {
-    Long save(User user);
-}
