--- src/main/java/com/example/scheduleapi/repository/user/UserRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.example.scheduleapi.repository.user;
-
-//public class UserRepositoryImpl implements UserRepository{
-//}
