+++ src/test/java/com/example/api/ScheduleApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.api;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ScheduleApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
