--- src/test/java/com/example/scheduleapi/ScheduleApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.scheduleapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ScheduleApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
